Fix from Jeremy Katz <katzj@redhat.com> for bug #169200
2005-03-04  Mark McLoughlin  <mark@skynet.ie>

        Fix from Jeremy Katz <katzj@redhat.com> for bug #169200

        * libmenu/desktop-entries.c (get_categories_from_key_file):
        sizeof(int) != sizeof(gsize).

+++ libmenu/desktop-entries.c
@@ -176,7 +176,7 @@ get_categories_from_key_file (DesktopEntry *entry,
 {
   GQuark  *retval;
   char   **strv;
-  int      len;
+  gsize    len;
   int      i;
 
   strv = g_key_file_get_string_list (key_file,
